Clean up deploy scripts and update initializer
